+++ src/events/DefineSourceSortOptionsEvent.php
@@ -15,7 +15,7 @@ use yii\base\Event;
  * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
  * @since 3.6.5
  */
-class DefineSourceSortOptionEvent extends Event
+class DefineSourceSortOptionsEvent extends Event
 {
     /**
      * @var string The element type class
@@ -28,16 +28,14 @@ class DefineSourceSortOptionEvent extends Event
     public $source;
 
     /**
-     * @var array|null The sort option definition.
+     * @var array The sort option definitions.
      *
-     * This should be set to an array with the following keys:
+     * Each sort option should be defined by an array with the following keys:
      *
      * - `label` – The sort option label
      * - `orderBy` – An array or comma-delimited string of columns to order the query by
      * - `attribute` _(optional)_ – The table attribute name that this option is associated with
      *   (required if `orderBy` is an array or more than one column name)
-     *
-     * Set to `null` to prevent the sort option from getting registered.
      */
-    public $sortOption;
+    public $sortOptions = [];
 }