+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ## Unreleased
 
 ### Added
-- Added `craft\events\DefineSourceSortOptionEvent`. ([#7515](https://github.com/craftcms/cms/issues/7515))
-- Added `craft\services\ElementIndexes::EVENT_DEFINE_SOURCE_SORT_OPTION`. ([#7515](https://github.com/craftcms/cms/issues/7515))
+- Added `craft\events\DefineSourceSortOptionsEvent`. ([#7515](https://github.com/craftcms/cms/issues/7515))
+- Added `craft\events\DefineSourceTableAttributesEvent`. ([#7515](https://github.com/craftcms/cms/issues/7515))
+- Added `craft\services\ElementIndexes::EVENT_DEFINE_SOURCE_SORT_OPTIONS`. ([#7515](https://github.com/craftcms/cms/issues/7515))
+- Added `craft\services\ElementIndexes::EVENT_DEFINE_SOURCE_TABLE_ATTRIBUTES`. ([#7515](https://github.com/craftcms/cms/issues/7515))
 
 ## 3.6.4.1 - 2021-02-05